% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_hessian.R
\name{dir.hessian}
\alias{dir.hessian}
\title{Hessian matrix of Dirichlet distribution with respect to parameter alpha}
\usage{
dir.hessian(x, mu, alpha, param = "alpha")
}
\arguments{
\item{x}{The data matrix of size $p \times n$ which is generated from \link{p}-dimensional Dirichlet distribution.}

\item{mu}{A positive scalar parameter which represents the dispersion parameter of the Dirichlet distribution.}

\item{alpha}{A $p \times 1$ parameter vector which is the log-transformed mean of Dirichlet distribution.}

\item{param}{A character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".
Off-diagonal elements of Hessian Matrix
Diagonal elements of Hessian Matrix}
}
\description{
Hessian matrix of Dirichlet distribution with respect to parameter alpha
}
