% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_jacobian.R
\name{dir.jacobian}
\alias{dir.jacobian}
\title{Jacobian of Dirichlet distribution with respect to parameter alpha}
\usage{
dir.jacobian(x, mu, alpha, param = "alpha")
}
\arguments{
\item{x}{A matrix size of (p x n) which is generated from \link{p}-dimensional Dirichlet distribution.}

\item{mu}{Scalar parameter of the Dirichlet distribution which represents the dispersion of the data.}

\item{alpha}{Vector of length p which represents the log-transformed mean vector of Dirichlet distribution.}

\item{param}{Character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".}
}
\value{
a Jacobian, the first-order partial derivatives with respect to param alpha, size of $(p-1)$ vector
}
\description{
Jacobian of Dirichlet distribution with respect to parameter alpha
}
