% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_lkhd.R
\name{dir.lkhd}
\alias{dir.lkhd}
\title{Dirichlet likelihood function}
\usage{
dir.lkhd(x, mu, alpha)
}
\arguments{
\item{x}{Data matrix of size \eqn{p \times n} generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{mu}{A scalar parameter which represents the dispersion parameter of the Dirichlet distribution.}

\item{alpha}{A vector of length \eqn{p-1} which is the mean parameter of Dirichlet distribution.}
}
\value{
a log-likelihood function of Dirichlet distribution.
}
\description{
Function to compute the Dirichlet likelihood function given a vector of observations and parameters of the distribution
}
