% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_lrt.R
\name{dir.lrt}
\alias{dir.lrt}
\title{Likelihood ratio test for Dirichlet distribution}
\usage{
dir.lrt(
  x,
  y = NULL,
  mu.x,
  mu.y = NULL,
  alpha.x,
  alpha.y = NULL,
  alpha,
  type = "one"
)
}
\arguments{
\item{x}{A data matrix of size \eqn{p \times n} generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{y}{A data matrix of size \eqn{p \times n} generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{mu.x}{A scalar parameter of x which represents the dispersion parameter of the Dirichlet distribution}

\item{mu.y}{A scalar parameter of y which represents the dispersion parameter of the Dirichlet distribution}

\item{alpha.x}{Vector of length p which represents the log-transformed mean vector of x.}

\item{alpha.y}{Vector of length p which represents the log-transformed mean vector of y.}

\item{alpha}{Vector of length p which represents the log-transformed mean vector under the null hypothesis.}

\item{type}{If \code{type} is  set as \code{one}, the function will return a one sample test result, else if \code{type} is  set as \code{two} then the function will return two samples test result.}
}
\value{
the function will return a list with two elements
\itemize{
\item test statistic,
\item p value
}
test statistic of LRT
}
\description{
Likelihood-ratio test for the parameter alpha of the Dirichlet distribution
}
