% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_hessian.R
\name{dm.hessian}
\alias{dm.hessian}
\title{Hessian matrix of Dirichlet distribution with respect to parameter alpha}
\usage{
dm.hessian(x, mu, alpha, param = "alpha")
}
\arguments{
\item{x}{The data matrix of size $(p x n)$ which is generated from \link{p}-dimensional Dirichlet distribution.}

\item{mu}{A positive scalar parameter which represents the dispersion parameter of the Dirichlet distribution.}

\item{alpha}{A $(p x 1)$ parameter vector which is the log-transformed mean of Dirichlet distribution.}

\item{param}{A character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".}
}
\value{
If \code{param} is set as \code{alpha}, the function will return a $(p-1) x (p-1)$ Hessian matrix, else if \code{param} is set as \code{mu}, the function will return a scalar of Hessian
Off-diagonal elements of Hessian Matrix
Diagonal elements of Hessian Matrix
}
\description{
Hessian matrix of Dirichlet distribution with respect to parameter alpha
}
