% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_jacobian.R
\name{dm.jacobian}
\alias{dm.jacobian}
\title{Jacobian for Dirichlet-multinomial}
\usage{
dm.jacobian(x, mu, alpha, param = "alpha")
}
\arguments{
\item{x}{A matrix size of \eqn{p \times n} which is generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{mu}{Scalar parameter of the Dirichlet distribution which represents the dispersion of the data.}

\item{alpha}{Vector of length p which represents the log-transformed mean vector of Dirichlet distribution.}

\item{param}{Character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".}
}
\value{
If \code{param} is set as \code{alpha}, the function will return a \eqn{(p-1) \times 1} vector of Jacobian, else if \code{param} is set as \code{mu}, the function will return a scalar of Jacobian.
Jacobian in terms of parameter alpha
Jacobian in terms of parameter mu
}
\description{
Jacobian of Dirichlet-multinomial distribution with respect to the two parameters: alpha and mu
}
