% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_lkhd.R
\name{dm.lkhd}
\alias{dm.lkhd}
\title{Dirichlet-multinomial likelihood function}
\usage{
dm.lkhd(x, mu, alpha)
}
\arguments{
\item{x}{Data matrix of size $p x n$ generated from \eqn{p}-dimensional Dirichlet-multinomial  distribution.}

\item{mu}{A scalar parameter which represents the dispersion parameter of the distribution.}

\item{alpha}{A vector of length \eqn{p - 1} which is the mean parameter of the distribution.}
}
\value{
a scalar which is the log-likelihood of the Dirichlet-multinomial distribution.
}
\description{
Function to compute the Dirichlet-multinomial distribution given a vector (or matrix) of observations and parameters of the distribution.
}
