% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_lrt.R
\name{dm.lrt}
\alias{dm.lrt}
\title{Likelihood ratio test for Dirichlet multinomial distribution}
\usage{
dm.lrt(
  x,
  y = NULL,
  mu.x,
  mu.y = NULL,
  alpha.x,
  alpha.y = NULL,
  alpha.null,
  type = "one"
)
}
\arguments{
\item{x}{A data matrix of size \eqn{p \times n} generated from \link{p}-dimensional Dirichlet-Multinomial distribution.}

\item{y}{A data matrix of size \eqn{p \times n} generated from \link{p}-dimensional Dirichlet-Multinomial distribution.}

\item{mu.x}{A scalar parameter of x which represents the dispersion parameter.}

\item{mu.y}{A scalar parameter of y which represents the dispersion parameter.}

\item{alpha.x}{Vector of length p which represents the log-transformed mean vector of x.}

\item{alpha.y}{Vector of length p which represents the log-transformed mean vector of y.}

\item{alpha.null}{Vector of length p which represents the log-transformed mean vector under the null hypothesis.}

\item{type}{If \code{type} is  set as \code{one}, the function will return a one sample test result, else if \code{type} is  set as \code{two} then the function will return two samples test result.}
}
\value{
the function will return test statistic, p value and the result of the Likelihood-ratio test.
test statistic of LRT
}
\description{
Likelihood-ratio test to parameter alpha of the Dirichlet-multinomial distribution
}
