% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_wald_test.R
\name{dm.wald}
\alias{dm.wald}
\title{Wald-type test to parameter alpha}
\usage{
dm.wald(
  x,
  y,
  mu.x,
  mu.y = NULL,
  alpha.null = NULL,
  alpha.x,
  alpha.y = NULL,
  type = "one"
)
}
\arguments{
\item{alpha.x}{Vector of length p which represents the log-transformed mean vector of x.}

\item{alpha.y}{Vector of length p which represents the log-transformed mean vector of y.}

\item{type}{If \code{type} is  set as \code{one}, the function will return a one sample test result, else if \code{type} is  set as \code{two} then the function will return two samples test result.}

\item{alpha}{Vector of length p which represents the log-transformed mean vector under the null hypothesis.}
}
\value{
the function will return test statistic, p value and the result of the Wald-type test.
}
\description{
Wald-type test to parameter alpha
}
