% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_randproj.R
\name{ortho.randproj}
\alias{ortho.randproj}
\title{Orthogonal random matrix generator}
\usage{
ortho.randproj(nrow, ncol, method = "norm", seed = NULL)
}
\arguments{
\item{nrow}{Number of rows in the random matrix to be generated}

\item{ncol}{Number of colums}

\item{method}{The method to be used for generating elements in the matrix. If \code{method = "norm"} (Default), elements are generated from standard normal distribution. If \code{method = "achlioptas"}, the elements are selected from the set \eqn{\{-1, 0, 1\}} with probabilities \eqn{\{1/6, 2/3, 1/6\}} respectively.}

\item{seed}{Set the seed to replicate the random matrix generated. Default is \code{NULL} and no seed is used.}
}
\value{
An orthogonal matrix of with \code{nrow} rows and \code{ncol} columns.
}
\description{
Generate random matrices of given dimension which have full row and are orthogonal
}
\examples{
a = ortho.randproj(nrow = 5, ncol = 3)
}
