% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_proj.R
\name{random.proj}
\alias{random.proj}
\title{Random matrix generator}
\usage{
random.proj(nrow, ncol, seed = NULL)
}
\arguments{
\item{nrow}{Number of rows in the random matrix to be generated}

\item{ncol}{Number of columns}

\item{seed}{Seed for random number generation. Default is NULL, where a random seed is used. However, the seed can be specified for reproducibility.}
}
\value{
A matrix of specified dimension with zero-one switch based random projections. The matrix is verified to have full row rank.
}
\description{
This function generates a random matrix whose elements are zeros and ones. The matrix is of full row rank.
}
