% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_jacobian.R
\name{dir.jacobian}
\alias{dir.jacobian}
\title{Jacobian for Dirichlet distribution}
\usage{
dir.jacobian(x, mu, alpha, param = "alpha")
}
\arguments{
\item{x}{A matrix size of \eqn{p \times n} which is generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{mu}{Scalar parameter of the Dirichlet distribution which represents the dispersion of the data.}

\item{alpha}{Vector of length p which represents the log-transformed mean vector of Dirichlet distribution.}

\item{param}{Character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".}
}
\value{
A Jacobian vector containing the first-order partial derivatives with respect to param alpha, of length \eqn{p-1}. For "mu", the function returns a scalar Jacobian.
}
\description{
This function computes the Jacobian of Dirichlet distribution with respect to the two parameters: alpha and mu
}
