% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_nr.R
\name{dir.nr}
\alias{dir.nr}
\title{Newton-Raphson for Dirichlet distriibution}
\usage{
dir.nr(x, mu = NULL, alpha = NULL, tol = 1e-04, maxiters = 100)
}
\arguments{
\item{x}{A data matrix of size \eqn{p \times n} generated from \eqn{p}-dimensional Dirichlet distribution.}

\item{mu}{A scalar parameter which represents the dispersion parameter of the Dirichlet distribution}

\item{alpha}{Vector of length p which represents the log-transformed mean vector of Dirichlet distribution.}

\item{tol}{Tolerance cutoff for determining convergence of the Newton-Raphson algorithm}

\item{maxiters}{Maximum number of iterations for the algorithm. Default is set at 100.}
}
\value{
the function will return a vector of length \eqn{p} containing the estimate of the param alpha and a the estimate of mu
}
\description{
Newton-Rapshon algorithm to estimate parameter alpha (or theta)
}
