% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_nr.R
\name{dm.nr}
\alias{dm.nr}
\title{Newton-Rapshon algorithm to estimate parameter alpha (or theta)}
\usage{
dm.nr(x, mu = NULL, alpha = NULL, tol = 1e-04, maxiters = 100)
}
\arguments{
\item{mu}{A scalar parameter which represents the dispersion parameter of the distribution}

\item{alpha}{Vector of length p which represents the log-transformed mean vector of the distribution.}

\item{tol}{Tolerance cutoff for determining convergence of the Newton-Raphson algorithm}

\item{maxiters}{Maximum number of iterations for the algorithm. Default is set at 100.}

\item{param}{Character string identifying the parameter for which the Jacobian is being computed. The possible values are "mu" or "alpha".}
}
\value{
the function will return a (p x 1) vector containing the estimate of the param alpha and a the estimate of mu
}
\description{
Newton-Rapshon algorithm to estimate parameter alpha (or theta)
}
